[MOV] stores: move table resize store in correct folder
Task: 6408800
