+++ src/components/tables/table_resizer/table_resizer.ts
@@ -1,7 +1,7 @@
 import { proxy, useProps } from "@odoo/owl";
 import { Component } from "../../../owl3_compatibility_layer";
-import { TableResizeStore } from "../../../plugins/ui_feature/table_resize_ui";
 import { useStore } from "../../../store_engine/store_hooks";
+import { TableResizeStore } from "../../../stores/table_resize_store";
 import { ViewportsStore } from "../../../stores/viewports_store";
 import { HeaderIndex, Highlight, Zone } from "../../../types/misc";
 import { SpreadsheetChildEnv } from "../../../types/spreadsheet_env";
